fixed new invoice page
git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@1732 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.BusinessLogic/InvoiceBL.cs
@@ -78,7 +78,7 @@ namespace HPF.FutureState.BusinessLogic
                 {
                     InvoiceCaseDTO invoiceCase = new InvoiceCaseDTO();
                     invoiceCase.InvoiceId = invoiceId;
-                    invoiceCase.InvoiceCaseBillAmount = (double)fCaseDraf.Amount;
+                    invoiceCase.InvoiceCaseBillAmount = fCaseDraf.Amount;
                     invoiceCase.ForeclosureCaseId = fCaseDraf.ForeclosureCaseId;
                     invoiceCase.ChangeLastAppName = invoiceDraft.ChangeLastAppName;
                     invoiceCase.ChangeLastDate = invoiceDraft.ChangeLastDate;
